Add skeleton __main__
